chore(backend): Refactor skills and occupations vector search COM-165, COM-166, COM-167, COM-106.
