+++ backend/app/vector_search/similarity_search_service.py
@@ -17,17 +17,7 @@ class SimilaritySearchService(ABC, Generic[T]):
 
         :param query: The query to search for.
         :param k: The number of results to return.
-        """
-        raise NotImplementedError
 
-    @abstractmethod
-    async def search_mmr(self, query: str, k: int = 5, fetch_k: int = 100, lambda_mult: float = 0.5) -> list[T]:
-        """
-        Perform a Maximal Marginal Relevance search on the vector store.
-
-        :param query: The query to search for.
-        :param k: The number of results to return.
-        :param fetch_k: The number of documents to pass to the MMR algorythm.
-        :param lambda_mult: The lambda multiplier for the MMR algorythm.
+        :return: A list of T objects.
         """
         raise NotImplementedError
