+++ backend/pyproject.toml
@@ -20,6 +20,7 @@ langchain-mongodb = "^0.1.3"
 deprecated="^1.2.14"
 google-cloud-dlp="^3.16.0"
 fix-busted-json=">=0.0.17,<1.0.0"
+motor = "^3.4.0"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^8.1.1"
